requirements.txt
Cython==3.0.5
numpy==1.26.2
sagemath==1.3.0
scipy==1.11.4
